Move contents of testing_test.go to two_pipelines_test.go
