+++ central/sensor/service/pipeline/nodeinventory/two_pipelines_test.go
@@ -22,6 +22,7 @@ import (
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/metrics"
 	nodeEnricher "github.com/stackrox/rox/pkg/nodes/enricher"
+	"github.com/stackrox/rox/pkg/scancomponent"
 	"github.com/stackrox/rox/pkg/scanners"
 	"github.com/stackrox/rox/pkg/scanners/types"
 	"github.com/stretchr/testify/assert"
@@ -368,4 +369,36 @@ func (*fakeNodeScanner) Type() string {
 
 func (*fakeNodeScanner) Name() string {
 	return "name"
+}
+
+func getDummyRisk() *storage.Risk {
+	return &storage.Risk{
+		Score:   1.0,
+		Results: make([]*storage.Risk_Result, 0),
+		Subject: &storage.RiskSubject{},
+	}
+}
+
+type mockNodeScorer struct{}
+
+func (m *mockNodeScorer) Score(_ context.Context, _ *storage.Node) *storage.Risk {
+	return getDummyRisk()
+}
+
+type mockComponentScorer struct{}
+
+func (m *mockComponentScorer) Score(_ context.Context, _ scancomponent.ScanComponent, _ string) *storage.Risk {
+	return getDummyRisk()
+}
+
+type mockDeploymentScorer struct{}
+
+func (m *mockDeploymentScorer) Score(_ context.Context, _ *storage.Deployment, _ []*storage.Risk) *storage.Risk {
+	return getDummyRisk()
+}
+
+type mockImageScorer struct{}
+
+func (m *mockImageScorer) Score(_ context.Context, _ *storage.Image) *storage.Risk {
+	return getDummyRisk()
 }
